Trivial Drive Billing
- Removed Unnecessary files.
- Added Build type for WS in DEV or PROD
- Added new variables
