--- app/src/main/java/com/aptoide/iabexample/Skus.java
@@ -1,16 +1,0 @@
-package com.aptoide.iabexample;
-
-/**
- * Created by neuro on 21-03-2018.
- */
-
-public final class Skus {
-
-  // SKUs for our products: the premium upgrade (non-consumable) and gas (consumable)
-  static final String SKU_PREMIUM_ID = "premium";
-  static final String SKU_GAS_ID = "gas";
-
-  // SKU for our subscription (infinite gas)
-  static final String SKU_INFINITE_GAS_MONTHLY_ID = "infinite_gas_monthly";
-  static final String SKU_INFINITE_GAS_YEARLY_ID = "infinite_gas_yearly";
-}
