+++ app/src/main/java/com/aptoide/iabexample/util/Skus.java
@@ -1,0 +1,16 @@
+package com.aptoide.iabexample.util;
+
+/**
+ * Created by neuro on 21-03-2018.
+ */
+
+public final class Skus {
+
+  // SKUs for our products: the premium upgrade (non-consumable) and gas (consumable)
+  public static final String SKU_PREMIUM_ID = "premium";
+  public static final String SKU_GAS_ID = "gas";
+
+  // SKU for our subscription (infinite gas)
+  public static final String SKU_INFINITE_GAS_MONTHLY_ID = "infinite_gas_monthly";
+  public static final String SKU_INFINITE_GAS_YEARLY_ID = "infinite_gas_yearly";
+}
